Create smear_pixel.py
You can use this function like this:

Copy code
smear_pixels("input.jpg", "output.jpg")
This code will take the image input.jpg smear it randomly and save the result as output.jpg. This is just an example and you can use this as a starting point and customize it according to your requirement.