scramble_pixels2.py
You can use this function like this:

Copy code
scramble_pixels("input.jpg", "output.jpg")
This code will take the image input.jpg and scramble the pixels in a pattern based on the golden ratio and save the result as output.jpg. The output image will be different each time you run this code as it is a random process.
This is just an example and you can use this as a starting point and customize it according to your requirement.