ascii_art.py
This script uses the Python Imaging Library (PIL) to open the image and convert it to grayscale. It then defines a list of ASCII characters to represent the different grayscale values in the image. The script then resizes the image to a smaller size (this step is optional) and gets the pixel data from the image. Finally, it creates an ASCII art string by iterating through the pixel data and adding the appropriate ASCII character for each pixel's grayscale value.

Please note that this is a basic example and doesn't include error handling or options to adjust the parameters of the image to get the best results. Also, this script will work with python 2.x and you may need to modify it accordingly if you are using python 3.x.