Create convertimage_hex.py
In this example, we open an image file in PNG format using the Image.open() method and pass the image object to BytesIO() method. Then we save the image in PNG format to the binary stream using image.save() method. After that, we get the binary data from the stream and convert it to a hexadecimal string using the hex() method. Finally, we print the hexadecimal string to verify the output.

This code assumes that the image file is located in the same directory as the Python script, you can change the path in the Image.open() method to load image from any location.

This code will work if the image is in the png format, if the image is in other format you should use appropriate method like jpeg or bmp to save the image.