Create verify_pgp_signature.py
This example will import the public key from the provided file, read the signature and data files and use the gpg.verify() method to check the signature. If the signature is valid, the script will print "Signature is valid!", otherwise it will print "Signature is invalid!".
It is also important to note that this script assumes that the key used to sign the data is present in the public_key_path file provided.

Also, you can use different PGP libraries like pycryptodomex, cryptography etc. to accomplish the same task.